requirements.txt
fastapi 
uvicorn 
openai 
requests 
aiofiles 
boto3 
redis 
pinecone
python-dotenv
pydantic-settings
google-cloud-storage
python-multipart
